refactor: 영화세부페이지 중복내용 MovieList 컴포넌트로 따로 생성
