+++ 호담-최규호/4주차/movie_detail_page_practice/src/assets/movie2.jpg
