+++ 호담-최규호/4주차/movie_detail_page_practice/src/components/Navbar.jsx
@@ -1,6 +1,6 @@
 import { Link } from "react-router-dom"
 import styled from "styled-components";
-import Logo from "../img/logo.png";
+import Logo from "../assets/logo.png";
 
 const TopBarColor = styled.div`
     display: flex;
