+++ 호담-최규호/4주차/movie_detail_page_practice/src/main.jsx
@@ -2,7 +2,7 @@ import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import App from './App.jsx'
 import './index.css'
-import logo from './img/logo.png';
+import logo from './assets/logo.png';
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
